docs: split DS→Multicenter into two scripts (one per reduction)
verify_ds_multicenter.py (combined) →
- verify_ds_minmax_multicenter.py (§4.1): 3,911 PASS
- verify_ds_minsum_multicenter.py (§4.2): 7,333 PASS

Now 9 scripts for 9 reductions, 1:1 correspondence.
MinSum script adds tight-bound check: non-DS always has
total distance > n-K (verifies the backward direction's
lower bound argument).
